refactored code added controllers for different routes
